GS-11244 refactor: Use barrel file to export hooks
